refactor(rust): split compiler into smaller files
